Input handling implemented
New Battlefield has a new attribute:
var battlefield = new Battlefield(canvas, width, height,
                                        tileSize, notifyFunc);
notifyFunc will be called whenever an attribute is changed.

For now it just can notify if the marked field is changed.  That happens
when the user klicks on the battlefield.
You can parse it like that:
function notifyFunc(state) {
    if (state.type == 'marked') {
        var posX = state.x;
        var posY = state.y;
        var fieldTypeName = state.name;
        /* Do some stuff here */
    }
}